Move `ArchiveUtils` from `core-utils` to `common-utils`
This class can be used independently of ORT. Note that `core-utils` still
needs to depend on `commons-compress` for its
`XZCompressedLocalFileStorage`.
