+++ .gitattributes
@@ -33,7 +33,7 @@ utils/spdx/src/main/resources/licenserefs/*	linguist-detectable=false
 utils/spdx/src/main/resources/licenses/*	linguist-detectable=false
 
 # Use Unix line endings for the expected control file of a Debian package.
-utils/core/src/test/assets/control-expected.txt  text eol=lf
+utils/common/src/test/assets/control-expected.txt  text eol=lf
 
 # Use Unix line endings for *.graphql files. This is mainly done to prevent that tests need to fiddle with
 # different platform-specific line endings.
