+++ downloader/src/funTest/kotlin/vcs/CvsWorkingTreeFunTest.kt
@@ -29,8 +29,8 @@ import java.io.File
 
 import org.ossreviewtoolkit.model.VcsInfo
 import org.ossreviewtoolkit.model.VcsType
+import org.ossreviewtoolkit.utils.common.unpack
 import org.ossreviewtoolkit.utils.core.ortDataDirectory
-import org.ossreviewtoolkit.utils.core.unpack
 import org.ossreviewtoolkit.utils.test.Ci
 import org.ossreviewtoolkit.utils.test.createSpecTempDir
