+++ downloader/src/funTest/kotlin/vcs/GitWorkingTreeFunTest.kt
@@ -31,8 +31,8 @@ import java.io.File
 import org.ossreviewtoolkit.downloader.VersionControlSystem
 import org.ossreviewtoolkit.model.VcsInfo
 import org.ossreviewtoolkit.model.VcsType
+import org.ossreviewtoolkit.utils.common.unpack
 import org.ossreviewtoolkit.utils.core.ortDataDirectory
-import org.ossreviewtoolkit.utils.core.unpack
 import org.ossreviewtoolkit.utils.test.createSpecTempDir
 
 class GitWorkingTreeFunTest : StringSpec({
