+++ scanner/src/main/kotlin/scanners/Askalono.kt
@@ -44,11 +44,11 @@ import org.ossreviewtoolkit.scanner.ScanException
 import org.ossreviewtoolkit.scanner.experimental.LocalScannerWrapper
 import org.ossreviewtoolkit.utils.common.Os
 import org.ossreviewtoolkit.utils.common.ProcessCapture
+import org.ossreviewtoolkit.utils.common.unpackZip
 import org.ossreviewtoolkit.utils.core.ORT_NAME
 import org.ossreviewtoolkit.utils.core.OkHttpClientHelper
 import org.ossreviewtoolkit.utils.core.createOrtTempDir
 import org.ossreviewtoolkit.utils.core.log
-import org.ossreviewtoolkit.utils.core.unpackZip
 import org.ossreviewtoolkit.utils.spdx.calculatePackageVerificationCode
 
 class Askalono(
