+++ utils/common/build.gradle.kts
@@ -17,6 +17,7 @@
  * License-Filename: LICENSE
  */
 
+val commonsCompressVersion: String by project
 val log4jApiKotlinVersion: String by project
 val jacksonVersion: String by project
 val semverVersion: String by project
@@ -32,6 +33,7 @@ dependencies {
     api("com.vdurmont:semver4j:$semverVersion")
     api("org.apache.logging.log4j:log4j-api-kotlin:$log4jApiKotlinVersion")
 
+    implementation("org.apache.commons:commons-compress:$commonsCompressVersion")
     implementation("org.jetbrains.kotlin:kotlin-reflect")
     implementation("org.springframework:spring-core:$springCoreVersion")
 }
